// src/app/types.ts
export interface Trainee {
  id: number;
  batch: string;
  status: string;
  traineeName: string;
  kana: string;
  traineeZip: string;
  traineeAddress: string;
  category: string;
  nationality: string;
  birthday: string;
  age: string;
  gender: string;
  period: string;
  stayLimit: string;
  cardNumber: string;
  passportLimit: string;
  passportNumber: string;
  certificateNumber: string;
  applyDate: string;
  certDate: string;
  entryDate: string;
  renewStartDate: string;
  assignDate: string;
  endDate: string;
  moveDate: string;
  returnDate: string;
  employmentReportDate: string;
  trainingStartDate: string;
  trainingEndDate: string;
  memo: string;
  phaseHistory: any[];
}

export interface Company {
  id: string;
  companyName: string;
  settlement: string;
  representative: string;
  jobType: string;
  zipCode: string;
  address: string;
  tel: string;
  joinedDate: string;
  employeeCount: string;
  acceptance: string;
  investmentCount: string;
  investmentAmount: string;
  investmentPayDate: string;
  corporateNumber: string;
  laborInsurance: string;
  employmentInsurance: string;
  implementationNumber: string;
  acceptanceDate: string;
  industryCategory: string;
  officeZip: string;
  officeAddress: string;
  responsiblePerson: string;
  instructor: string;
  lifeInstructor: string;
  planInstructor: string;
  memo: string;
  trainees: Trainee[];
  createdAt?: any;
}